requirements.txt
paho-mqtt==1.6.1
protobuf==4.25.1
google-generativeai==0.3.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
# OAuth 2.0 and Gmail API support
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
